New: Add logic for deleting quota suspension periods

+++ app/helpers/workbasket_helper.rb
@@ -331,6 +331,11 @@ module WorkbasketHelper
           step: :main
         )
 
+    when :delete_quota_suspension
+      edit_delete_quota_suspension_url(
+        workbasket.id,
+        step: :main
+      )
       end
   end
 
@@ -374,6 +379,8 @@ module WorkbasketHelper
       create_quota_association_url(workbasket.id)
     when :delete_quota_association
       delete_quota_association_url(workbasket.id)
+    when :delete_quota_suspension
+      delete_quota_suspension_url(workbasket.id)
     when :create_quota_suspension
       create_quota_suspension_url(workbasket.id)
     when :edit_quota_suspension
@@ -420,6 +427,8 @@ module WorkbasketHelper
       withdraw_workbasket_from_workflow_create_quota_suspension_url(workbasket.id)
     elsif workbasket.object.type == "delete_quota_association"
       withdraw_workbasket_from_workflow_delete_quota_association_url(workbasket.id)
+    elsif workbasket.object.type == "delete_quota_suspension"
+      withdraw_workbasket_from_workflow_delete_quota_suspension_url(workbasket.id)
     end
   end
